pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rss-inbox"
version = "0.1.0"
description = "RSS feed processing and inbox management CLI tool"
authors = [
    {name = "User", email = "user@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]
keywords = ["rss", "feed", "inbox", "cli", "devops"]
dependencies = [
    "typer>=0.9.0",
    "feedparser>=6.0.0", 
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "urllib3>=2.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/user/rss-inbox"
Repository = "https://github.com/user/rss-inbox"
Issues = "https://github.com/user/rss-inbox/issues"

[project.scripts]
rss-inbox = "rss_inbox.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rss_inbox"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/CHANGELOG.md",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true